+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - audit logging functions require context
+- add functions to set environment, hostname and pod name for audit logs
 
 ### Deprecated